# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BrokerConnectionMeta": {
        "description": "Current state of a broker connection for a specific portfolio.",
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "broker_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broker Name"
          },
          "connection_state": {
            "default": "disconnected",
            "title": "Connection State",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "portfolio_id": {
            "title": "Portfolio Id",
            "type": "integer"
          },
          "sync_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Error"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "portfolio_id"
        ],
        "title": "BrokerConnectionMeta",
        "type": "object"
      },
...
      "ConnectRequest": {
        "description": "Body for POST /brokers/{portfolio_id}/connect.\nNon-secret fields only \u2014 the UI should never send API keys over this endpoint\nuntil secure credential handling (KMS / vault) is implemented.",
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Broker account reference (non-secret)",
            "title": "Account Id"
          },
          "broker_name": {
            "description": "e.g. 'zerodha' | 'ibkr'",
            "title": "Broker Name",
            "type": "string"
          },
          "config": {
            "additionalProperties": true,
            "description": "Non-secret config fields",
            "title": "Config",
            "type": "object"
          }
        },
        "required": [
          "broker_name"
        ],
        "title": "ConnectRequest",
        "type": "object"
      },
      "ConnectResponse": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "broker_name": {
            "title": "Broker Name",
            "type": "string"
          },
          "connection_state": {
            "title": "Connection State",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "portfolio_id": {
            "title": "Portfolio Id",
            "type": "integer"
          },
          "scaffolded": {
            "default": false,
            "title": "Scaffolded",
            "type": "boolean"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "portfolio_id",
          "broker_name",
          "connection_state",
          "message"
        ],
        "title": "ConnectResponse",
        "type": "object"
      },
...
      "DisconnectResponse": {
        "properties": {
          "broker_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broker Name"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "portfolio_id": {
            "title": "Portfolio Id",
            "type": "integer"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "portfolio_id",
          "message"
        ],
        "title": "DisconnectResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SyncResponse": {
        "properties": {
          "broker_name": {
            "title": "Broker Name",
            "type": "string"
          },
          "holdings_synced": {
            "default": 0,
            "title": "Holdings Synced",
            "type": "integer"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "portfolio_id": {
            "title": "Portfolio Id",
            "type": "integer"
          },
          "post_snap_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Snap Id"
          },
          "pre_snap_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pre Snap Id"
          },
          "rows_skipped": {
            "default": 0,
            "title": "Rows Skipped",
            "type": "integer"
          },
          "scaffolded": {
            "default": false,
            "title": "Scaffolded",
            "type": "boolean"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "portfolio_id",
          "broker_name",
          "message"
        ],
        "title": "SyncResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "P-Insight Portfolio Analytics API. Visit /docs for interactive API documentation.",
    "title": "P-Insight API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/brokers/{portfolio_id}/connect": {
      "post": {
        "description": "Initiate a broker connection for a portfolio.\n\nFor scaffolded connectors: persists a pending state and returns\n`scaffolded: true` with an explanatory message \u2014 no error raised.\n\nFor implemented connectors: calls the connector's auth flow and\nupdates the portfolio's source metadata.\n\nSecurity note: this endpoint accepts only non-secret config fields.\nAPI keys and tokens must be passed via environment variables until\nsecure credential storage is implemented.",
        "operationId": "connect_broker_api_v1_brokers__portfolio_id__connect_post",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connect a broker account to a portfolio",
        "tags": [
          "Brokers"
        ]
      }
    },
    "/api/v1/brokers/{portfolio_id}/connection": {
      "delete": {
        "description": "Remove the broker connection for a portfolio.",
        "operationId": "disconnect_broker_api_v1_brokers__portfolio_id__connection_delete",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisconnectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disconnect a broker from a portfolio",
        "tags": [
          "Brokers"
        ]
      },
      "get": {
        "description": "Returns the current broker connection state for a given portfolio.\nIf no connection exists, returns `connection_state: \"disconnected\"`.",
        "operationId": "get_connection_api_v1_brokers__portfolio_id__connection_get",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrokerConnectionMeta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get broker connection state for a portfolio",
        "tags": [
          "Brokers"
        ]
      }
    },
    "/api/v1/brokers/{portfolio_id}/sync": {
      "post": {
        "description": "Pull current holdings from the connected broker and replace\nthe portfolio's holdings (with pre/post snapshots for history).\n\nReturns HTTP 501 with `scaffolded: true` for unimplemented connectors.\nReturns HTTP 400 if no connection exists or state is not connected.",
        "operationId": "sync_broker_api_v1_brokers__portfolio_id__sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync holdings from a broker into a portfolio",
        "tags": [
          "Brokers"
        ]
      }
    },
...
    "/api/v1/news/": {
      "get": {
        "description": "Return recent news articles relevant to portfolio holdings.\n\n- If tickers is provided, only articles matching at least one of those\n  tickers are returned.\n- If tickers is omitted, all holdings are fetched and used as the filter.\n- event_type further narrows the result set.\n\nPhase 1: Served from mock_data/portfolio.json (static).\nPhase 2: Fetched from live news API by ticker.",
        "operationId": "get_news_api_v1_news__get",
        "parameters": [
          {
            "description": "Comma-separated ticker list to filter articles (e.g. TCS.NS,INFY.NS). If omitted, all portfolio holdings are used.",
            "in": "query",
            "name": "tickers",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated ticker list to filter articles (e.g. TCS.NS,INFY.NS). If omitted, all portfolio holdings are used.",
              "title": "Tickers"
            }
          },
          {
            "description": "Filter by event type. One of: earnings, dividend, deal, rating, company_update, market_event, regulatory, management",
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type. One of: earnings, dividend, deal, rating, company_update, market_event, regulatory, management",
              "title": "Event Type"
            }
          },
          {
            "description": "Data source mode",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "uploaded",
              "description": "Data source mode",
              "enum": [
                "uploaded",
                "live",
                "broker"
              ],
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get portfolio-relevant news",
        "tags": [
          "News & Events"
        ]
      }
    },
    "/api/v1/news/events": {
      "get": {
        "description": "Return upcoming corporate events (earnings dates, dividends, AGMs).\n\nEvents are sorted soonest-first.\nPhase 1: Static mock events. Phase 2: Live corporate calendar API.",
        "operationId": "get_events_api_v1_news_events_get",
        "parameters": [
          {
            "description": "Comma-separated ticker list. If omitted, all holdings are used.",
            "in": "query",
            "name": "tickers",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated ticker list. If omitted, all holdings are used.",
              "title": "Tickers"
            }
          },
          {
            "description": "Filter by event type (earnings, dividend, agm, bonus, split).",
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type (earnings, dividend, agm, bonus, split).",
              "title": "Event Type"
            }
          },
          {
            "description": "Data source mode",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "uploaded",
              "description": "Data source mode",
              "enum": [
                "uploaded",
                "live",
                "broker"
              ],
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get upcoming corporate events",
        "tags": [
          "News & Events"
        ]
      }
    },
...
  }
}